Fix restart for Zed (#7486)
Quick change to fix restart for Zed.

The issue is that Zed is not sending expected restart sequence, which
includes the `Exited` => `Starting` state transition. When this state
transition does not occur, the Console doesn't get a signal that the
restarted runtime is coming back online. Since Zed currently does
`Exited` => `Initializing` => `Starting`, the console did not detect
that it had restarted.

Addresses https://github.com/posit-dev/positron/issues/7432.