Assistant: Improve Execute Code ergonomics (#7496)
This change improves the Execute Code tool in Positron Assistant in a
couple of key areas (detailed below).

Addresses https://github.com/posit-dev/positron/issues/7040.

### Toolbar

The Execute Code tool now shows a code toolbar. This allows you to copy
or paste the code if running it wasn't what you wanted.

<img width="321" alt="image"
src="https://github.com/user-attachments/assets/3c4e2029-aa75-49ff-b35a-6c10647e4592"
/>

### Persistence

When the code is run, it doesn't disappear but instead stays in the chat
history (several users pointed out that the disappearing code was
confusing). You can even see the status while it's running, and a green
check when it's done.

<img width="313" alt="image"
src="https://github.com/user-attachments/assets/7ab38daa-50e9-4c66-b484-d8c0a0d15462"
/>

### Implementation

This turned out to be a lot easier than I thought, because it turns out
that this problem was largely solved already upstream: lots of special
casing exists for Terminal tool invocations, for all the same reasons we
wanted to special case Console tool invocations. So most of this change
is simply making our Execute Code tool look enough like a Terminal tool
that we get the same UI.

### QA Notes

Note that the "Run" button in the toolbar is not present on Execute Code
blocks even though it _is_ present on regular code blocks. This is
intentional because it's confusing to have two run gestures (toolbar +
Continue button), and it would be difficult to make the toolbar button
do the same thing as the Continue button.

The issue #7040 mentions streaming responses, but that isn't covered in
this PR. I think it's less important and can be done separately.