Rework console interrupt so that Ctrl-C and the interrupt action bar button do the same thing (#7495)
This PR addresses the remaining issue in #697 so that pressing Ctrl-C and using the interrupt action bar button behave exactly the same and interrupt the console, whether there is an active prompt or not. Additionally, after interrupt, focus is now driven into the code editor widget. Previously, focus went (seemingly) nowhere when interrupting a prompt.