Supervisor: Fix problem starting sessions after a long idle period (#7475)
This change addresses an issue that can happen when Positron is left
running for a long time with no sessions running.

The underlying issue is in Kallichore. It is designed to shut down after
a time if Positron is not connected to it, as a safeguard to avoid
orphaned supervisor processes littering the system. However, it uses
active websocket connections to determine whether Positron is connected.
If you have no sessions running and therefore no websockets connected,
the server will clean itself up after some time (default is an hour).

If you then try to start a new session after the server has exited, an
inscrutable `AggregateError` is shown.

The fix has two layers:
- Most importantly, I've added a client heartbeat endpoint to the
server, which Positron now pings every 20 seconds, whether there are
sessions running or not. This prevents the server from shutting itself
down when Positron is still running.
- Secondly, if you do attempt to start a session when the server is not
running, a single attempt is made to start the server and retry the
session start.

Two other minor improvements are included:
- The full error is now shown instead of the inscrutable
`AggregateError`. The full error isn't a lot more helpful, but it is at
least more specific. (More work would be helpful here.)
- The console now treats this kind of startup failure correctly (i.e.
shows it as exited instead of still starting).

See also https://github.com/posit-dev/kallichore/issues/22. 

Addresses https://github.com/posit-dev/positron/issues/7438.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fix `AggregateError` starting an interpreter if Positron is left idle
for more than an hour with no interpreters running
(https://github.com/posit-dev/positron/issues/7438).


### QA Notes