Fix IntelliSense for viewsContainers.auxiliarybar contributions (#7513)
Since #2895 it's been possible (though undocumented) for Positron
extensions to contribute views containers to the sidebar, unlike
upstream VS Code.

This worked because upstream used to ignore additional keys in
`contributes.viewsContainers`, but since
https://github.com/microsoft/vscode/pull/241700 they've explicitly
prohibited them, so IntelliSense is starting to flag these keys as
errors.

This commit fixes this by explicitly adding
`contributes.viewsContainers.auxiliarybar` as a key in the JSON schema.

Addresses #7512.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>