Data Explorer: Preserve non-file URIs when creating DuckDB clients (#7533)
Previously we encoded only the original path in the data explorer,
losing the scheme (and other URI attributes) and fundamentally assuming
that we could only open local files.

This commit instead encodes the original URI in its entirety and adds a
helper to retrieve it, which makes it possible to use the `Open as Plain
Text` button even with files backed by a virtual filesystem provider.

Double-encoding the URI here is pretty gross, of course, but it's
consistent with what we were already doing.

This is the last piece of #7351.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>