Hide Chat panel when Assistant is disabled (#7519)
This change hides the Chat panel entirely when Positron Assistant is
disabled. The approach taken is:

- Never show the Chat panel unless there are chat agents registered. The
upstream configuration shows the panel in other contexts, such as when
an incorrect version of Copilot is installed -- not something we need to
worry about since we bundle Assistant.
- Do not register any chat agents when Assistant is disabled. This is
necessary since the default agent comes from a built-in extension and is
therefore always registered.

Addresses https://github.com/posit-dev/positron/issues/6828. 

### QA Notes

A reload or restart is required to apply the setting.