Add posit jenkins bot to CLA allowlist (#7609)
I created a GitHub action to automatically update Positron and PWB Code
server with the latest Workbench extension number when the extension
successfully builds. This is set up and working and has created its
first PR: https://github.com/posit-dev/positron/pull/7607

That PR is currently blocked because the bot hasn't signed the CLA. This
PR adds the bot to our allowlist.