Fix `tar-fs` vulnerability in all positron extensions (#7600)
Partially addresses
[#349](https://github.com/posit-dev/positron-builds/issues/349)

This PR fixes a security vulnerability in the tar-fs package by updating
the dependency that depended on it ( `@vscode/vsce`) to the latest
version.

Next steps will be to fix bad dependencies upstream.




### Release Notes

Fix security vulnerability in `tar-fs` dependency in positron extensions

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

Everything should continue to run as previously... hopefully.