Rewrite file hyperlinks on web (#7621)
Addresses #7256

File hyperlinks in the Console should now open the targetted file, in a
web/workbench context.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- File hyperlinks in the R Console now open the relevant remote file
when running on Workbench.

### QA Notes

@:web @:win @:console

The test that is being un-skipped here is a good manual test case. In an
instance of Positron (Web), execute something like this:

``` r
library(cli)
cli_text("Let's open {.file DESCRIPTION:15:8}")
```

where `DESCRIPTION` is a path to a file that exists and the `:15` or
`:15:8` part is an optional specification of line or line and column.

---------

Signed-off-by: sharon <sharon-wang@users.noreply.github.com>
Co-authored-by: sharon wang <sharon-wang@users.noreply.github.com>