Bump ark to 0.1.183 (#7682)
- https://github.com/posit-dev/ark/pull/793
- https://github.com/posit-dev/ark/pull/794
- https://github.com/posit-dev/ark/pull/795
- https://github.com/posit-dev/ark/pull/796
- https://github.com/posit-dev/ark/pull/797
- https://github.com/posit-dev/ark/pull/798



### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- In curly braces, code is no longer indented before a minimum level
(#1683).
- Evaluating expressions in the console no longer causes focus to jump
back to the debug location (#3151).
- In debug mode, the call stack is now laid out in a more intuitive
manner (#5155).
- Set the `positron.show_last_value` global option to `TRUE` to show
`.Last.value` in the variables pane (#3034).
- Fixed an issue preventing evaluation of multi-line code in the
debugger (#5928).


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->