set model-specific `maxOutputTokens` so that larger text selections can be edited (#7702)
### Summary

- addresses #7500 and maybe also #7675
- set a larger max output token limit for each model, based on the
documented max output token limits for each model
- set the default limit to 4096 (only used for the Echo and Error test
models at the moment), which is the max token output for Claude 3 Opus
and Claude 3 Haiku (seemed like a reasonable default 🤷)
- note that it can take a long time to get suggested edits, but it does
work now. We may need to do some form of paging of the content to
achieve something a bit faster.

### Release Notes

#### Bug Fixes

- Positron Assistant: Increased the maximum token output to allow for
edits to larger selections of text (#7500, #7675)

### QA Notes

@:assistant

The issue was consistently reproducible for me when using
<kbd>Cmd+I</kbd> to invoke inline assistant edits, but only sometimes
reproducible when using the Chat in both Ask and Edit modes.

Note that it is not the chat message that needs to be long, but the
selection of text/code that should be quite long (2500+ characters seems
to do the trick!).

1. Select 2500+ characters in a file
2. <kbd>Cmd+I</kbd> for inline assistant
3. Ask the assistant to modify all of the text in some way
4. It may take a moment, but should succeed and apply an edit.
Previously, it would run and say it did something, but no edits would be
provided.