Refine Assistant follow-ups prompt (#7708)
The existing prompt may have led models on tangents by specifying "data
science questions or directions". @hadley proposed this streamlined
prompt that does seem to reduce the number of irrelevant follow-ups when
they are suggested. We can monitor and continue to refine these going
forward.

Addresses #7637

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- N/A

### QA Notes

Let's continue to monitor this and work on an evaluation test for the
quality of this prompt.