Add support for images in tool call results for vercel-anthropic, take 2 (#7717)
This addresses #7705. The purpose of this change is for a modified
version of the `executeCode` tool, which can return a result object that
contains both text and images.

When Anthropic is used via the Vercel back end, it can now send images
in tool call results. Also:

- When this is used, it is no longer necessary to special-case handling
of `getPlot` tool call results, where it replaces the tool call message
and the tool result message with normal Assistant and User messages.
- This can handle tool results where the `content` includes both text
and image parts.

This PR does not affect providers other than Anthropic.

Note that this also probably should be used in conjunction with the
base64 padding fix mentioned here:

https://github.com/posit-dev/positron/issues/7510#issuecomment-2882149754

I didn't know whether or not that should go in a separate PR, so I left
it out for now.