Assistant: remove assistant guide link (#7806)
The _Positron Assistant User Guide_ link doesn't go to an actual user
guide, and the consensus from triage was that we will just remove it
until we have somewhere for it to go. It's just as will since the
[welcome screen is too tall as it
is](https://www.github.com/posit-dev/positron/issues/7580) and this
saves some space.

Addresses https://github.com/posit-dev/positron/issues/7628.