Remove "experimental" remote SSH setting (#7805)
Removes the "experimental" remote SSH setting. This setting hid the
action in the bottom left of the window that initiates a remote
connection. Now that it's been removed, it will always be available.

<img width="208" alt="image"
src="https://github.com/user-attachments/assets/d7a3fefb-63d8-4889-ba24-95d1d9e895f6"
/>

Addresses #7581.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- The Remote Status Indicator in the bottom left, formerly behind an
Experimental setting, is now always available (#7581)

#### Bug Fixes

- N/A


### QA Notes

This new behavior removes Positron overrides and restores base Code OSS
functionality.