assistant: replace empty tool results with placeholder text (#7799)
### Summary

- addresses https://github.com/posit-dev/positron/issues/7770
- improves message processing to handle empty content more effectively
    - we should not filter out _all_ empty responses
- we can continue to filter out messages that have no content at all and
messages that only contain empty text parts
- however, we should not filter out tool results with empty content
altogether
- now, when a tool result part has empty content, we include placeholder
content in place of the empty content, to ensure that tool calls will
have corresponding tool results
- updates our workspace tools to return non-empty messages
    - throw an error when the tool criteria is not met
- added text to tool descriptions to note that empty workspaces are not
supported
- updates anthropic integration test to cover these cases
- adds whether workspace folders are open to the general context, so
that LLMs are less likely to use the workspace tools when the workspace
criteria is not met

### Release Notes

#### Bug Fixes

- Assistant: ensure tool calls have corresponding tool results (#7770)

### QA Notes

@:assistant

The easiest test case is running the project tree tool, file contents
tool or text search tool when in a workspace-less window. The Assistant
should not call any of these tools since no workspace folders are open.
If you are able to sidestep this and get the Assistant to call the tools
anyways, the tool should now throw a more user-friendly error, directing
the user to ensure that a workspace folder is open.