Fix "Apply in Editor" tool on Windows (#7822)
This change is the minimum needed to make _Apply in Editor_ work again.

In earlier iterations of the internal API, there was an
`editing-session` agent location (corresponding to the Copilot Edits
pane). This location doesn't exist any more; there's just a Chat pane.

The fix here is to register the agent formerly used as the editing
session agent for the "Edit" mode instead of trying to use the default
agent for both Ask and Edit modes.

Addresses #7188.