E2E test: only show import vscode settings popup when needed for related test (#7840)
Added:

"positron.importSettings.enable": false

to .vscode/settings.json in qa-example-content to prevent the "Import
VSCode Settings" dialog from appearing in all tests. Also need to turn
in back on for the "Import VSCode Settings" tests. This PR does that and
changes the scope of the setting to allow for the qa-example-content
setting to work in the first place.

### QA Notes

@:vscode-settings @:win