Update "Interpreter" command palette actions (#7803)
Addresses #7189 

This PR goes through all the interpreter commands we expose to users in
the command palette and updates them so the naming scheme is consistent
and works better with multiple interpreter sessions. This also updates a
couple other UI element labels that trigger the command palette actions
so labels match across the board.

Note: The diff is a bit messy looking because I reordered a bunch of the
actions, so hiding whitespace is recommended. I am also commenting the
actions that were renamed/removed so its easier to find the destructive
changes.

For most of the actions, a user is now prompted to select an interpreter
session they want to take the action on. There are a couple actions that
specifically act on the foreground session/active console, and those now
specify that by using the term "Active Interpreter Session/ Active
Console".

There’s a few actions in this list that I wasn’t sure what to do with
and would love feedback on:
* Create/Close Runtime Client Widget: This action gets exposed in the
command palette and I wanted to ask if this should be shown to users. I
don’t know what the use case is for this. Alternatively, do we want to
move this into the “Developer” category?
* Execute Code in Console/Execute Code Silently: from the comments it
looks like this command is mainly used by QA. Can this be moved into the
“Developer” category? I don’t love providing this as a path for users to
run code because it’s a pretty odd experience.

Here's the new list of interpreter actions after the cleanup:

<img width="621" alt="Screenshot 2025-05-21 at 2 47 41 PM"
src="https://github.com/user-attachments/assets/cfc18351-4cf1-4fa1-8071-6330a9090f04"
/>

The plus button in the console pane was also updated so the hover label
matches the action being triggered. Prior to the fix, the button always.
showed the "Duplicate Session" label - even when the action being
triggered was the create new session action.


https://github.com/user-attachments/assets/29d5c067-e3de-4452-919b-0be833206ec8

### Release Notes

#### New Features

- Updated "Interpreter" command palette actions to work better with
multiple interpreter sessions.
- Fixed incorrect hover label for "+" button in console pane

#### Bug Fixes

- N/A


### QA Notes

@:sessions @:web @:win @:console

---------

Signed-off-by: Dhruvi Sompura <dhruvi.sompura@posit.co>
Co-authored-by: Julia Silge <julia.silge@gmail.com>
Co-authored-by: Marie Idleman <marie.idleman@posit.co>