update secrets baseline and ensure hooks fail precommit (#7855)
For some reason, the detect-secrets and detect-dot-only hooks have
stopped failing the precommit hook.

The changes in this PR ensure that the hooks exit the process with an
error, which prevents a commit from succeeding.

The secrets baseline file was also missing a false positive mark on an
item -- that has now been updated.