App Launcher: handle python app terminal links in `positron-run-app` extension (#7847)
### Summary

- addresses #5431

#### Implementation

- we now register a TerminalLinkProvider in the positron-run-app
extension which handles the python app urls in the Terminal, instead of
going through the usual uri opener flow (which does not have access to
the proxied app urls)
- api.ts: we now keep track of the app servers and corresponding
terminals so that we can handle the terminal links and resolve them to
the correct proxied app url
- reorganized constants, types, api-utils into separate files to reduce
the size of the main app launcher api.ts file
- we don't proxy web apps through the positron-proxy when running on
Desktop anymore -- this seemed unnecessary 🤷

#### Demo


https://github.com/user-attachments/assets/7350d6f4-2b9c-442a-b422-8eab797fccc8

#### `positron.appLauncher.terminalAppUrlOpenLocation` setting

This setting allows the user to control where the app url will open when
Cmd/Ctrl-clicked in the Terminal. A gear icon shows in the default
quickpick that shows when the user Cmd/Ctrl-clicks the terminal link
which navigates the user to the setting in the UI.

### Release Notes

#### Bug Fixes

- App Launcher: app urls from Terminal links are now resolved correctly
(#5431)

### QA Notes

@:web @:apps @:win

This was most obviously broken for Dash apps on Web, however all
supported App Launcher app types should work.

1. Run a supported App Launcher Python app in the Terminal via the play
button
2. Note the url that displays in the Viewer when the app is loaded
3. Cmd/Ctrl+click the url of the app in the Terminal
4. A quickpick should open and the url in the quickpick placeholder area
should match the URL in the Viewer
5. Selecting either of the options (viewer pane or browser window)
should load the app successfully