Queue operations on plot comms per session (#7858)
Prior to this change, each plot maintained its own render queue, with
the result that when multiple plots were emitted, they would all try to
render at once, resulting in timeouts. This change serializes all of the
plot rendering work per session, so that:

- Render requests are not sent to the session until the session is idle.
- Only one render request is sent at a time. 

Addresses https://github.com/posit-dev/positron/issues/5954. 

Some dev notes:

- The reason things worked as they did in the past is that we didn't
want to presume that a render would be handled on the main thread
(perhaps some plots or some languages could render plots in the
background). However, practically speaking this is the case for both our
R and Python kernels. If we want to get maximum performance from
backends that can do background rendering, we'll have to unwind some
parts of this change and/or make it possible for the runtime to declare
whether it can handle parallel or background plot rendering.
- This is a targeted change for plot comms, but practically speaking
almost every comm could benefit from a task queue/serializer so that
RPCs are executed in order and do not overlap with each other or with
RPCs from other comms connected to the same runtime. We may want to
consider making this a generic part of the comm infrastructure when we
have the capacity for more risk.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Fix timeouts when rendering plots during slow-running operations or in
a loop (#5954)

### QA Notes

The behavior is better for R than Python, since R sends a pre-render of
each plot while code is running. For Python, you can expect that you'll
have to wait to see the plot until the session is idle. This work is
tracked in #7451.

Test tags: `@:plots`