Expose Positron Assistant settings (#7871)
Exposes Positron Assistant settings in its package.json so that they are
available in VS Code's Settings UI.

Addresses https://github.com/posit-dev/positron/issues/7852.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Positron Assistant, an experimental AI assistant for Positron, can now
be enabled in Settings.
https://github.com/posit-dev/positron/issues/7852

#### Bug Fixes

- N/A


### QA Notes

No logic has changed here.