Ignore exited status of session in console dropdown
It is not common for a session and its console instance
to hang around when the session has a state of exited.
This happens when something goes awry with the
session startup/restart/exit. In this situation, providing
the runtime for the user is helpful so they can start up
another session of the same runtime.