Fix issue starting R/Python when `/bin/sh` is user's shell (#7881)
This change addresses an issue in which starting R/Python sessions fails
when `/bin/sh` is the user's default shell.

Vanilla `sh` does not support the long form `--login` flag, only `-l` to
run login scripts. Thankfully bash-alikes support `-l` as an alias to
`--login` too, so we can support both with `-l`.

(realize that there are a long tail of other shells we can probably test
with like `fish`, `csh`/`tcsh`, etc; the intent of this fix is to just
cover the most common shells for now, which is `bash`, `zsh`, and `sh`)

Addresses https://github.com/posit-dev/positron/issues/7874.

### QA Notes

- See the original issue for instructions on changing your shell to
`/bin/sh` to test this.
- Before verifying the issue, I recommend running `echo $SHELL` in a
Terminal to verify that you are really using `sh` as a shell.