Fix showing local web servers over Remote SSH (#7921)
This change fixes an issue that happens in Remote SSH environments when
viewing content from web servers on the remote machine.

<img width="1691" alt="image"
src="https://github.com/user-attachments/assets/eb23f363-876b-47e3-ba5a-95657c43c99a"
/>

The issue was somewhat intermittent because it's timing-related. Very
briefly, content from web servers on Remote SSH is made available on the
client machine by creating an SSH tunnel that forwards the port, e.g.
from `localhost:1234` on the remote machine to `localhost:1234` on the
client. However, we load the URL immediately in the Viewer, which causes
it to attempt to load before the tunnel is created in some cases.

The fix is very straightforward: when running in a remote environment,
we add the `allowTunneling` flag when resolving the URL. This causes the
port forwarding to happen immediately (before we attempt to load the
URL), and also has the advantage of correcting the URL in the case that
there is a port conflict.

Addresses https://github.com/posit-dev/positron/issues/7406.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fix failure loading local web server (e.g. Shiny) content in the
Viewer pane in Remote SSH sessions (#7406)


### QA Notes